moving and renaming xdsml project helper
- removed the reference to Melange in the name (there is no dependency
to melange)
- move to an utils plugin
